Add files via upload
This was another project of creating a tribute page to someone/something. I followed FCC for the formatting of the page but played around with a few ideas to try to get what I wanted. This is a project I will likely go back to quite a few times because I thought it was pretty fun and I would like to really flesh out every aspect of it from the background to the formatting around the picture and the text sizes/fonts.